#201 Add the toc to Layout
