+++ src/components/layout/ContentBox.tsx
