+++ src/components/layout/defaultLayout.module.scss
@@ -1,0 +1,8 @@
+.mainPane {
+    display: flex;
+}
+
+.leftPane {
+    position: sticky;
+    top: 0;
+}
